Create Part_1_worksheet.md
Except for exercise 4 and adding "justice" and "fairness" to the list, this is the worksheet I used at adacamp.

My original exercise 4 was the question: what is the project or accomplishment you're proud of that you've been the most afraid to "own" publicly.  I didn't word it exactly that way, and a lot of folks interpreted it as being "what is your worst fear" - which was not really the positive note I wanted to end on :(

So this exercise 4 is a lot simpler; I'll move the "proudest secret" one to Part 2.